+++ CHANGELOG.md
@@ -8,7 +8,16 @@ benefit enrollment opportunities are available to an employee.
 
 ### Next release: Putatively 7.1.4
 
++ Prefer to source manager "Time/Absence Dashboard" URL from new HRS URLs
+  service key `Time/Absence Dashboard`, falling back on now-deprecated portlet
+  preference for this URL. [HRSPLT-454][]
 + Remove "View Time Entry Exceptions" link. [HRSPLT-453][]
+
+#### Deprecated in 7.1.4
+
++ Time and Absence portlet preference `approvalsDashboardUrl` is deprecated.
+  Provision the time and absence approvals dashboard URL via new HRS URLs DAO
+  key `Time/Absence Dashboard`.
 
 ### 7.1.3 Really fix Benefit Information widget "Learn more" link
 
@@ -1050,3 +1059,4 @@ This and many more earlier releases exist as [releases in the GitHub repo][].
 [HRSPLT-437]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-437
 [HRSPLT-449]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-449
 [HRSPLT-453]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-453
+[HRSPLT-454]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-454